chore: fix flow configuration to reduce errors
Including the `react-navigation/TypeDefinition.js` in the flow config
`[libs]` section led to `react-navigation` not being able to find the
type defitions itself - which created over 200 errors.
By removing the file from the `[libs]` section we can reduce the flow
errors to about 90 errors, most of them in the application code.