chore: configurable environment
In order to be able to configure values through environment variables
we want to add a "./scripts/configure.sh" which can be used to generate
a "app/config.js" file with values taken from the environment.
This script generates the config file and therefore needs to be run
before every build / environment switch.