chore: check whether the keystore variables are set before using it
After building a freshly cloned version of this repository the build
step fails because the gradle global variables for the release key store
are not set.
This commit adds a check around the signing configs to test whether
these variables are available and only if the are available sets up the
signing config.